Add desktop application for Go-DNS
This commit introduces a new file, desktopApp.go, which contains code for a GUI-based desktop application for the Go-DNS project. The application provides features to check the DNS MX, TXT and CNAME entries for a given domain. The GUI is created using the Fyne library and has fields for domain input and the corresponding DNS record results. The application is designed to make the record lookup process more user-friendly.