Add SSL checker feature and readme file update
Added a new SSL checking feature in the 'sslchecker.go' file. This new function will allow users to check the SSL status of a given domain. Added SSL Checker in the desktop app to call the SSL checking function from the application. README file updated with new feature description and usage. Changes made to enhance the functionality of the application by providing an additional tool for users to check network safety.