Update README.md with packaging instructions
Updated README.md with instructions on how to package the Go-DNS app using the Fyne toolkit. Removed the development related lines and replaced it with a link to the release page of the project. Also added a new section "Packaging" to give more detailed instructions for setting up Fyne and packaging the app.